+++ mencoder.c
@@ -163,8 +163,8 @@ static float audio_delay=0.0;
 static int ignore_start=0;
 static int audio_density=2;
 
-float force_fps=0;
-static float force_ofps=0; // set to 24 for inverse telecine
+double force_fps=0;
+static double force_ofps=0; // set to 24 for inverse telecine
 static int skip_limit=-1;
 float playback_speed=1.0;
