+++ mplayer.c
@@ -302,7 +302,7 @@ static int ignore_start=0;
 
 static int softsleep=0;
 
-       float force_fps=0;
+       double force_fps=0;
 static int force_srate=0;
 static int audio_output_format=-1; // AF_FORMAT_UNKNOWN
        int frame_dropping=0; // option  0=no drop  1= drop vo  2= drop decode
