+++ mplayer.h
@@ -22,7 +22,7 @@ extern char * font_name;
 extern char * sub_font_name;
 extern float  font_factor;
 extern float movie_aspect;
-extern float force_fps;
+extern double force_fps;
 
 //extern char **sub_name;
 extern float  sub_delay;
